+++ spec/defines/aem_instance_spec.rb
@@ -5,7 +5,8 @@ describe 'aem::instance', :type => :defines do
 
   let :default_facts do
     {
-      :kernel => 'Linux'
+      :kernel           => 'Linux',
+      :operatingsystem  => 'CentOS',
     }
   end
 
@@ -55,6 +56,16 @@ describe 'aem::instance', :type => :defines do
     end
 
     it do
+      is_expected.to contain_aem__service('aem').with(
+        'ensure' => 'present',
+        'status' => 'enabled',
+        'home'   => '/opt/aem',
+        'user'   => 'aem',
+        'group'  => 'aem'
+      )
+    end
+
+    it do
       is_expected.to contain_aem__config(
         'aem'
       ).with(
@@ -93,9 +104,12 @@ describe 'aem::instance', :type => :defines do
     it { is_expected.to contain_group('aem').that_requires('Anchor[aem::aem::begin]') }
     it { is_expected.to contain_user('aem').that_requires('Anchor[aem::aem::begin]') }
 
+    # This doesn't work, so there's no test case for the dependency tree.
+    #it { is_expected.to contain_aem__service('aem').that_requires('Aem::License <| |>') }
     it { is_expected.to contain_aem_installer('aem').that_requires('Aem::Config[aem]') }
     it { is_expected.to contain_aem__config('aem').that_requires('Aem::Package[aem]') }
     it { is_expected.to contain_aem__package('aem').that_requires('Anchor[aem::aem::begin]') }
+
 
     it do
       is_expected.to contain_file(
@@ -176,6 +190,10 @@ describe 'aem::instance', :type => :defines do
     it { is_expected.to contain_user('aem').with('ensure' => 'absent') }
     it { is_expected.to contain_group('aem').with('ensure' => 'absent') }
 
-    it { is_expected.to contain_aem__package('aem').that_requires('Anchor[aem::aem::begin]') }
+    it { is_expected.to contain_aem__package('aem').with('ensure' => 'absent') }
+    it { is_expected.to contain_aem__service('aem').with('ensure' => 'absent') }
+
+    it { is_expected.to contain_aem__service('aem').that_requires('Anchor[aem::aem::begin]') }
+    it { is_expected.to contain_aem__package('aem').that_requires('Aem::Service[aem]') }
   end
 end